+++ src/app/app-routing.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LoginMainComponent } from './login-main/login-main.component';
-import { RegisterMainComponent } from './register-main/register-main.component';
-import { ViewMainComponent } from './view-main/view-main.component';
-import { JiraUsersConfigComponent } from './jira-users-config/jira-users-config.component';
+import { LoginMainComponent } from './components/login-main/login-main.component';
+import { RegisterMainComponent } from './components/register-main/register-main.component';
+import { ViewMainComponent } from './components/view-main/view-main.component';
+import { JiraUsersConfigComponent } from './components/jira-users-config/jira-users-config.component';
 import { RouteGuardsGuard as AuthGuard } from './guards/route-guards.guard';
-import { NotfoundComponent } from './notfound/notfound.component';
+import { NotfoundComponent } from './components/notfound/notfound.component';
 import { CantAccessGuard } from './guards/cant-access.guard';
-import { UploadFileComponent } from './upload-file/upload-file.component';
-import { DetailCertificateComponent } from './detail-certificate/detail-certificate.component';
+import { UploadFileComponent } from './components/upload-file/upload-file.component';
+import { DetailCertificateComponent } from './components/detail-certificate/detail-certificate.component';
 
 const routes: Routes = [
   { path: 'login', component: LoginMainComponent },
