+++ src/app/components/detail-certificate/detail-certificate.component.spec.ts
