+++ src/app/components/detail-certificate/detail-certificate.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Certificate } from '../models.interface';
-import { ApiService } from '../services/api.service';
+import { Certificate } from '../../models.interface';
+import { ApiService } from '../../services/api.service';
 
 @Component({
   selector: 'app-detail-certificate',
