+++ src/app/components/index.ts
@@ -1,0 +1,16 @@
+import { DetailCertificateComponent } from "./detail-certificate/detail-certificate.component";
+import { JiraUsersConfigComponent } from "./jira-users-config/jira-users-config.component";
+import { LoginMainComponent } from "./login-main/login-main.component";
+import { NavbarComponent } from "./navbar/navbar.component";
+import { NavbarLoginComponent } from "./navbar-login/navbar-login.component";
+import { NotfoundComponent } from "./notfound/notfound.component";
+import { RegisterMainComponent } from "./register-main/register-main.component";
+import { ShowCertificatesComponent } from "./show-certificates/show-certificates.component";
+import { UploadFileComponent } from "./upload-file/upload-file.component";
+import { ViewMainComponent } from "./view-main/view-main.component";
+
+export {
+    DetailCertificateComponent, JiraUsersConfigComponent, LoginMainComponent,
+    NavbarComponent, NavbarLoginComponent, NotfoundComponent, RegisterMainComponent,
+    ShowCertificatesComponent, UploadFileComponent, ViewMainComponent
+};
