+++ src/app/components/navbar-login/navbar-login.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { AuthenticatedService } from '../services/authenticated.service';
-import { ApiService } from '../services/api.service';
+import { AuthenticatedService } from '../../services/authenticated.service';
+import { ApiService } from '../../services/api.service';
 
 @Component({
   selector: 'app-navbar-login',
