+++ src/app/components/register-main/register-main.component.spec.ts
