+++ src/app/components/view-main/view-main.component.css
